Extract shared TextWidgetAnswer
