+++ collect_app/src/main/java/org/odk/collect/android/widgets/WidgetAnswer.kt
@@ -1,13 +1,19 @@
 package org.odk.collect.android.widgets
 
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AttachFile
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
 import androidx.lifecycle.ViewModelProvider
 import org.javarosa.core.model.Constants
 import org.javarosa.form.api.FormEntryPrompt
-import org.odk.collect.android.widgets.arbitraryfile.ArbitraryFileWidgetAnswer
-import org.odk.collect.android.widgets.barcode.BarcodeWidgetAnswer
+import org.odk.collect.android.widgets.arbitraryfile.ArbitraryFileWidgetAnswerViewModel
 import org.odk.collect.android.widgets.video.VideoWidgetAnswer
+import org.odk.collect.icons.R
 
 @Composable
 fun WidgetAnswer(
@@ -22,8 +28,9 @@ fun WidgetAnswer(
         when (prompt.controlType) {
             Constants.CONTROL_INPUT -> {
                 when (prompt.dataType) {
-                    Constants.DATATYPE_BARCODE -> BarcodeWidgetAnswer(
+                    Constants.DATATYPE_BARCODE -> TextWidgetAnswer(
                         modifier,
+                        ImageVector.vectorResource(R.drawable.ic_baseline_barcode_scanner_white_24),
                         answer,
                         fontSize,
                         onLongClick
@@ -31,13 +38,19 @@ fun WidgetAnswer(
                 }
             }
             Constants.CONTROL_VIDEO_CAPTURE -> VideoWidgetAnswer(modifier, answer, viewModelProvider!!, onLongClick)
-            Constants.CONTROL_FILE_CAPTURE -> ArbitraryFileWidgetAnswer(
-                modifier,
-                answer,
-                fontSize,
-                viewModelProvider!!,
-                onLongClick
-            )
+            Constants.CONTROL_FILE_CAPTURE -> {
+                val context = LocalContext.current
+                val viewModel = viewModelProvider!![ArbitraryFileWidgetAnswerViewModel::class]
+
+                TextWidgetAnswer(
+                    modifier,
+                    Icons.Default.AttachFile,
+                    answer,
+                    fontSize,
+                    onLongClick,
+                    stringResource(org.odk.collect.strings.R.string.open_file)
+                ) { viewModel.openFile(context, answer) }
+            }
             else -> throw IllegalArgumentException("Unsupported control type: ${prompt.controlType}")
         }
     }
